refactor(app): extract cwd tracking
